moving some code around and tweaking Aqua testing.
