+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ReferenceFiniteElements"
 uuid = "6dc62d09-f8eb-43fd-9672-074e490a997f"
 authors = ["Craig M. Hamel <cmhamel32@gmail.com> and contributors"]
-version = "0.7.2"
+version = "0.7.3"
 
 [deps]
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
@@ -22,13 +22,22 @@ ReferenceFiniteElementsAdaptExt = ["Adapt"]
 ReferenceFiniteElementsExodusExt = ["Exodus"]
 
 [compat]
+Adapt = "3"
+Aqua = "0.8"
+Distributions = "0.25"
 DocStringExtensions = "0.9"
+Exodus = "0.10"
 FastGaussQuadrature = "0.5, 1"
+ForwardDiff = "0.10"
+JET = "0.8"
+LinearAlgebra = "1.9"
 Polynomials = "3, 4"
 PrecompileTools = "1"
 SpecialPolynomials = "0.4"
 StaticArrays = "1"
 StructArrays = "0.6"
+Test = "1.9"
+TestSetExtensions = "2"
 julia = "1"
 
 [extras]
